Fix Infinite Loop/CORS Rigged for Dev/Malformed Import Statements
Had an issue where making any request from the home page caused the
browser to enter an infinite loop, eventually causing it to lock up.

CORS was having issues, so I set it up for a dev environment and the
error coupled with a network error indicated that the api server wasn't running.

The API server had malformed import statements.